Added some documentation, improved methods. Added some new methods as well to aid in IEconService.

+++ CSharpTradeOffers/Trading/IEconServiceHandler.cs
@@ -7,6 +7,9 @@ using Newtonsoft.Json;
 
 namespace CSharpTradeOffers.Trading
 {
+    /// <summary>
+    /// Handles EconService related tasks such as trade offers.
+    /// </summary>
     public class EconServiceHandler
     {
         //create initliazer to pass inventoryhandler?
@@ -17,6 +20,10 @@ namespace CSharpTradeOffers.Trading
         private const string ApiUrl = "https://api.steampowered.com/IEconService/{0}/v1/"; //migrate to base url later
         private const string BaseUrl = "https://api.steampowered.com/IEconService/";
 
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
+        /// <param name="apiKey"></param>
         public EconServiceHandler(string apiKey)
         {
             _apiKey = apiKey;
@@ -25,9 +32,8 @@ namespace CSharpTradeOffers.Trading
         /// <summary>
         /// Requests all trade offers.
         /// </summary>
-        /// <param name="_apiKey">The bot's API key.</param>
         /// <param name="data">A dictionary containing the URL params found here: https://developer.valvesoftware.com/wiki/Steam_Web_API/IEconService at GetTradeOffers (v1).
-        /// ex: <"get_sent_offers","1"></param>
+        /// ex: [leftarrow]"get_sent_offers","1"[rightarrow]. Please note, [leftarrow] and [rightarrow] correspond to the keys on the keyboard.</param>
         /// <returns></returns>
         public TradeOffersList GetTradeOffers(Dictionary<string, string> data)
         {
@@ -41,7 +47,6 @@ namespace CSharpTradeOffers.Trading
         /// <summary>
         /// Requests a single trade offer based on the tradeofferid.
         /// </summary>
-        /// <param name="_apiKey">The bot's API key.</param>
         /// <param name="tradeofferid">The tradeofferid to request information on.</param>
         /// <param name="language">The language to use. Default: english</param>
         /// <returns>A CEConTradeOffer object.</returns>
@@ -64,7 +69,6 @@ namespace CSharpTradeOffers.Trading
         /// <summary>
         /// Declines a trade offer that was sent to you.
         /// </summary>
-        /// <param name="_apiKey">The bot's API key.</param>
         /// <param name="tradeofferid">The ID of the offer to decline.</param>
         /// <returns></returns>
         public string DeclineTradeOffer(string tradeofferid)
@@ -81,7 +85,6 @@ namespace CSharpTradeOffers.Trading
         /// <summary>
         /// Cancels a trade offer based on the specified tradeofferid.
         /// </summary>
-        /// <param name="_apiKey">The bot's API Key.</param>
         /// <param name="tradeofferid">The ID of the offer to cancel.</param>
         /// <returns></returns>
         public string CancelTradeOffer(string tradeofferid)
@@ -185,7 +188,6 @@ namespace CSharpTradeOffers.Trading
         /// <summary>
         /// Attempts to locate a valid trade in the TradeConfig.
         /// </summary>
-        /// <param name="_apiKey">The bot's API Key.</param>
         /// <param name="offer">A TradeOffer object containing the trade to create a compatible form with.</param>
         /// <param name="inventoryHandler">An InventoryHandler object, the InventoryHandler must be that off the bot otherwise it will return erroneous data.</param>
         /// <returns>A compatible trade offer, null if none was found.</returns>
@@ -286,6 +288,11 @@ namespace CSharpTradeOffers.Trading
             return null;
         }
 
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// IGNORE THIS METHOD FOR NOW, IT IS BAD!
+        /// </summary>
+        /// <param name="assets"></param>
         public void GetNames(ref List<CEconAsset> assets)
         {
             foreach (CEconAsset cEconAsset in assets)
@@ -450,14 +457,38 @@ namespace CSharpTradeOffers.Trading
 
 
     #region enums
+    /// <summary>
+    /// TypeIds for use in trade config, mostly for reference and not actually used anywhere
+    /// </summary>
     public enum TypeIds
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         Exact = 0,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         Contains = 1,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         StartsWith = 2,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ClassId = 3,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         DollarWorth = 4,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         Tag = 5,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         CurrencyTrade = 6
     }
 
@@ -569,71 +600,190 @@ namespace CSharpTradeOffers.Trading
         PhoneActivityLimitExceeded = 97
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     public enum ETradeOfferState
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateInvalid = 1,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateActive = 2,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateAccepted = 3,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateCountered = 4,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateExpired = 5,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateCanceled = 6,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateDeclined = 7,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateInvalidItems = 8,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateEmailPending = 9,
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         ETradeOfferStateEmailCanceled = 10
     }
 
     #endregion
 
     #region GetTradeOffersSummaryResponse
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "RootObject")]
     public class GetTradeOffersSummaryBaseResponse
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public GetTradeOffersSummaryResponse response { get; set; }
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "Response")]
     public class GetTradeOffersSummaryResponse
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int pending_received_count { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int new_received_count { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int updated_received_count { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int historical_received_count { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int pending_sent_count { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int newly_accepted_sent_count { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int updated_sent_count { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int historical_sent_count { get; set; }
     }
     #endregion
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "RootObject")]
     public class MarketValue
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public bool success { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string lowest_price { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string volume { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string median_price { get; set; }
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "RootObject")]
     public class TradeId
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public ulong tradeid { get; set; }
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "ItemsToReceive")]
     public class CEconAsset
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string appid { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string contextid { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string assetid { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string classid { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string instanceid { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string amount { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public bool missing { get; set; }
 
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         [JsonIgnore] public string name;
 
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
+        /// <param name="apiKey"></param>
+        /// <returns></returns>
         public string GetMarketHashName(string apiKey)
         {
             var _handler = new ISteamEconomyHandler();
@@ -643,63 +793,165 @@ namespace CSharpTradeOffers.Trading
         }
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "TradeOffersReceieved")]
     public class CEconTradeOffer
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string tradeofferid { get; set; }
-        public UInt32 accountid_other { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
+        public uint accountid_other { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string message { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int expiration_time { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int trade_offer_state { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public List<CEconAsset> items_to_give { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public List<CEconAsset> items_to_receive { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public bool is_our_offer { get; set; }
-        public int time_created { get; set; }
-        public int time_updated { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
+        public ulong time_created { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
+        public ulong time_updated { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string tradeid { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public bool from_real_time_trade { get; set; }
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "Response")]
     public class TradeOffersList
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public List<CEconTradeOffer> trade_offers_received { get; set; }
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "RootObject")]
     public class TradeOffers
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public TradeOffersList response { get; set; }
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "RootObject")]
     public class SendOfferResponse
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string tradeofferid { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public bool needs_email_confirmation { get; set; }
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public string email_domain { get; set; }
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     public class Me
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public List<CEconAsset> assets = new List<CEconAsset>();
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public List<object> currency = new List<object>();
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public bool ready = false;
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     public class Them
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public List<CEconAsset> assets = new List<CEconAsset>();
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public List<object> currency = new List<object>();
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public bool ready = false;
     }
 
+    /// <summary>
+    /// I forgot or it's obvious. TODO: Add better documentation
+    /// </summary>
     [JsonObject(Title = "RootObject")]
     public class TradeOffer
     {
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public bool newversion = true;
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public int version = 2;
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public Me me = new Me();
+        /// <summary>
+        /// I forgot or it's obvious. TODO: Add better documentation
+        /// </summary>
         public Them them = new Them();
     }
 }
